Małe poprawki wyświetlania danych poziomu natężenia światła

+++ src/sensorData.hpp
@@ -16,7 +16,7 @@ RainStatusSensor rainStatusSensor(2);
 
 float temperature;
 bool rainStatus;
-int lightLevel;
+unsigned long lightLevel;
 float pressure;
 tmElements_t tm;
